Force explicit int cast for decoct in PHP8
In PHP8 at least, the lack of cast was causing large values to be read as strings, causing errors such as this one: https://github.com/owncloud/TarStreamer/issues/23#issuecomment-1449700060

In my case, this happened when a large file had to be added to the Tar archive. Forcing the (int) cast seems to have fixed it for me, and the files are downloaded correctly. I don't expect any issues with PHP7 either, but please review if it's acceptable or if there's any other area where this would fit best.

Thanks.